--- react-app/src/features/faq/content/onboardingGuides.ts
@@ -1,24 +1,0 @@
-import { Template } from "../faqs/utils";
-
-export const ONBOARDING_GUIDES: Template[] = [
-  {
-    title: "Welcome to OneMAC",
-    href: "/onboarding/WelcometoOneMAC.pdf",
-  },
-  {
-    title: "IDM Instructions for OneMAC Users",
-    href: "/onboarding/IDMInstructionsforOneMACUsers.pdf",
-  },
-  {
-    title: "OneMAC IDM Guide",
-    href: "/onboarding/OneMACIDMGuide.pdf",
-  },
-  {
-    title: "OneMAC State User Guide",
-    href: "/onboarding/OneMACStateUserGuide.pdf",
-  },
-  {
-    title: "OneMAC CMS User Guide",
-    href: "/onboarding/OneMACCMSUserGuide.pdf",
-  },
-];
