+++ react-app/src/features/faq/faqs/general/browsers.tsx
@@ -1,0 +1,16 @@
+export const Browsers = () => (
+  <p>
+    The submission portal works best on Google Chrome (Version 91.0.4472.77 or later) or Firefox
+    (Version 89.0 or later). Firefox users may need to adjust their browser settings to enable PDFs
+    to download automatically —{" "}
+    <a
+      href="https://support.mozilla.org/en-US/kb/change-firefox-behavior-when-open-file"
+      rel="noopener noreferrer"
+      target="_blank"
+      className="text-primary underline hover:no-underline"
+    >
+      learn how
+    </a>
+    .
+  </p>
+);
