+++ react-app/src/features/faq/faqs/general/is-official.tsx
@@ -1,0 +1,10 @@
+export const IsOfficial = () => (
+  <p>
+    Yes, as long as you have the electronic receipt (confirmation email). Your submission is
+    considered your official state submission and will only be considered received by CMS if you
+    have received the electronic receipt. You should receive an email confirmation that the formal
+    action was received along with information about the 90th day. If you do not receive a
+    confirmation email for your SPA or waiver submissions, please contact your state lead or your
+    state’s CMS lead for HCBS or managed care.
+  </p>
+);
