+++ react-app/src/features/faq/faqs/general/onemac-roles.tsx
@@ -1,0 +1,13 @@
+import { PdfLink } from "../utils";
+
+export const OneMacRoles = () => (
+  <p>
+    Please refer to the State role descriptions that can be found on page 4 of the{" "}
+    <PdfLink
+      href="/onboarding/OneMACStateUserGuide.pdf"
+      label="general"
+      title="OneMAC State User Guide"
+    />
+    .
+  </p>
+);
