+++ react-app/src/features/faq/faqs/spa/chip-spa-attachments.tsx
@@ -1,0 +1,143 @@
+export const ChipSpaAttachments = ({
+  isChipSpaDetailsEnabled,
+}: {
+  isChipSpaDetailsEnabled: boolean;
+}) => (
+  <>
+    {isChipSpaDetailsEnabled && <p className="font-bold">CHIP SPA Attachment Types:</p>}
+    <p className="ml-12">Note: “*” indicates a required attachment.</p>
+    <table className="faq-table ml-12 border-collapse border border-gray-300 w-full ">
+      <tbody>
+        <tr>
+          <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+          <th className="border border-gray-300 px-4 py-2">Description</th>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Current State Plan*</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Current version of the CHIP state plan that details how the State operates its CHIP
+            program
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
+            Amended State Plan Language*
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            Track changes to <span className="underline">only</span> the currently approved CHIP
+            state plan pages that the State is proposing to amend
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Cover Letter*</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Cover letter to CMS with an authorized signature that outlines the purpose of the CHIP
+            SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Updated 1-year budget if applicable of the State's planned expenditures if the CHIP SPA
+            submission has a significant impact on the approved budget
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Public Notice</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Process used by the State if applicable to accomplish involvement of the public that
+            occurred specifically for this CHIP SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Consultation process with Indian Tribes if applicable that occurred specifically for
+            this CHIP SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Other</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Other supporting document(s) needed to process the CHIP SPA submission
+          </td>
+        </tr>
+      </tbody>
+    </table>
+
+    {isChipSpaDetailsEnabled && (
+      <>
+        <p className="font-bold mt-12">CHIP Eligibility SPA Attachment Types:</p>
+        <p className="ml-12">Note: “*” indicates a required attachment.</p>
+        <table className="faq-eligibility-table ml-12 border-collapse border border-gray-300 w-full ">
+          <tbody>
+            <tr>
+              <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+              <th className="border border-gray-300 px-4 py-2">Description</th>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
+                CHIP Eligibility Template*
+              </td>
+              <td className="border border-gray-300 px-4 py-2">
+                Amendment to the eligibility section of the CHIP state plan using a PDF template
+                repository, including statutory and regulatory background, required information, and
+                minimum review criteria
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Cover Letter*</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Cover letter to CMS with an authorized signature that outlines the purpose of the
+                CHIP SPA submission
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Current State Plan</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Current version of the CHIP state plan that details how the State operates its CHIP
+                program
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2 whitespace-nowrap">
+                Amended State Plan Language
+              </td>
+              <td className="border border-gray-300 px-4 py-2">
+                Track changes to <span className="underline">only</span> the currently approved CHIP
+                state plan pages that the State is proposing to amend
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Updated 1-year budget if applicable of the State's planned expenditures if the CHIP
+                SPA submission has a significant impact on the approved budget
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Public Notice</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Process used by the State if applicable to accomplish involvement of the public that
+                occurred specifically for this CHIP SPA submission
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Consultation process with Indian Tribes if applicable that occurred specifically for
+                this CHIP SPA submission
+              </td>
+            </tr>
+            <tr>
+              <td className="border border-gray-300 px-4 py-2">Other</td>
+              <td className="border border-gray-300 px-4 py-2">
+                Other supporting document(s) needed to process the CHIP Eligibility SPA submission
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </>
+    )}
+  </>
+);
