+++ react-app/src/features/faq/faqs/spa/chip-spa-rai-attachments.tsx
@@ -1,0 +1,55 @@
+export const ChipSpaRaiAttachments = () => (
+  <>
+    <p>Note: “*” indicates a required attachment.</p>
+    <table className="faq-table  border-collapse border border-gray-300 w-full">
+      <thead>
+        <tr>
+          <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+          <th className="border border-gray-300 px-4 py-2">Description</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Revised Amended State Plan Language*</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Revision made to the amended state plan language of the CHIP SPA submission. State to
+            provide a redline version and a clean version of the revised amended state plan pages
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Official RAI Response*</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Official response to CMS to support RAI inquiries for the CHIP SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Budget Documents</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Updated 1-year budget if applicable of the State’s planned expenditures if the CHIP SPA
+            submission has a significant impact on the approved budget
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Public Notice</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Process used by the State if applicable to accomplish involvement of the public that
+            occurred specifically for this CHIP SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Tribal Consultation</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Consultation process with Indian Tribes if applicable that occurred specifically for
+            this CHIP SPA submission
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Other</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Other document(s) needed to process the CHIP SPA submission
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </>
+);
