+++ react-app/src/features/faq/faqs/spa/mpc-spa-templates.tsx
@@ -1,4 +1,4 @@
-import { Template } from "../faqs/utils";
+import { PdfList, Template } from "../utils";
 
 export const MPC_TEMPLATES: Template[] = [
   {
@@ -31,3 +31,14 @@ export const MPC_TEMPLATES: Template[] = [
     href: "/mpc/G3.pdf",
   },
 ];
+
+export const MpcSpaTemplates = () => (
+  <section className="space-y-2">
+    <p>
+      Medicaid Premiums and Cost Sharing SPA templates can be downloaded at the links below. After
+      downloading and completing the templates you need, upload them as part of the SPA submission.
+      The template PDFs can only be opened using Adobe Reader or Acrobat.
+    </p>
+    <PdfList list={MPC_TEMPLATES} label="template" />
+  </section>
+);
