+++ react-app/src/features/faq/faqs/spa/public-health-emergency.tsx
@@ -1,0 +1,6 @@
+export const PublicHealthEmergency = () => (
+  <p>
+    Yes, all PHE-related SPAs should be submitted through OneMAC by completing the Medicaid SPA
+    form.
+  </p>
+);
