+++ react-app/src/features/faq/faqs/waiver/appk-attachments.tsx
@@ -1,0 +1,34 @@
+export const AppkAttachments = () => (
+  <>
+    <p>
+      The regulations at 42 C.F.R. §430.25, 431.55 and 42 C.F.R. §441.301 describe the requirements
+      for submitting section 1915(b) and 1915(c) waivers.
+    </p>
+    <p>Note: “*” indicates a required attachment.</p>
+    <table className="faq-table border-collapse border border-gray-300 w-full">
+      <thead>
+        <tr>
+          <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+          <th className="border border-gray-300 px-4 py-2">Description</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            1915(c) Appendix K Amendment Waiver Template*
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            Official amendments to 1915(c) waiver programs addressing or in response to Disasters or
+            Emergencies.
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Other</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Supplemental documents for the 1915(c) Appendix K waiver amendment
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </>
+);
