+++ react-app/src/features/faq/faqs/waiver/appk.tsx
@@ -1,0 +1,3 @@
+export const Appk = () => (
+  <p>Yes, you can submit Appendix K amendments in the 1915(c) Appendix K form.</p>
+);
