+++ react-app/src/features/faq/faqs/waiver/temporary-extensions-c-attachments.tsx
@@ -1,0 +1,28 @@
+export const TemporaryExtensionsCAttachments = () => (
+  <>
+    <p>Note: “*” indicates a required attachment.</p>
+    <table className="faq-table border-collapse border border-gray-300 w-full ">
+      <thead>
+        <tr>
+          <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+          <th className="border border-gray-300 px-4 py-2">Description</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Waiver Extension Request*</td>
+          <td className="border border-gray-300 px-4 py-2">
+            A formal letter addressed to George Failla, Director of the Division of HCBS Operations
+            & Oversight requesting a temporary extension beyond the current approved waiver period.
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Other</td>
+          <td className="border border-gray-300 px-4 py-2">
+            Supplemental documents for the Waiver Extension Request
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </>
+);
