+++ react-app/src/features/faq/faqs/waiver/waiver-extension-status.tsx
@@ -1,0 +1,7 @@
+export const WaiverExtensionStatus = () => (
+  <p>
+    Temporary Extensions Requests will only show a status of ‘Submitted’ in the OneMAC system at
+    this time. Their status does not update regardless of where that request is in the Submission
+    Review process.
+  </p>
+);
