+++ react-app/src/features/faq/faqs/waiver/waiver-id-help.tsx
@@ -1,0 +1,15 @@
+import { handleSupportLinkClick } from "../utils";
+
+export const WaiverIdHelp = () => (
+  <p>
+    Email{" "}
+    <a
+      className="text-primary"
+      href="mailto:MCOGDMCOActions@cms.hhs.gov"
+      onClick={handleSupportLinkClick("general")}
+    >
+      MCOGDMCOActions@cms.hhs.gov
+    </a>{" "}
+    to get support with determining the correct 1915(b) Waiver Number.
+  </p>
+);
