+++ react-app/src/features/faq/faqs/waiver/waiverb-attachments.tsx
@@ -1,0 +1,74 @@
+export const WaiverBAttachments = () => (
+  <>
+    <p>
+      The regulations at 42 C.F.R. §430.25, 431.55 and 42 C.F.R. §441.301 describe the requirements
+      for submitting section 1915(b) and 1915(c) waivers.
+    </p>
+    <table className="faq-table border-collapse border border-gray-300 w-full">
+      <thead>
+        <tr>
+          <th className="border border-gray-300 px-4 py-2">Attachment Name</th>
+          <th className="border border-gray-300 px-4 py-2">Description</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            1915(b)(4) FFS Selective Contracting (Streamlined) waiver application pre-print
+            (Initial, Renewal, Amendment)
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            State submission of the 1915(b)(4) Waiver Fee-for-Service Selective Contracting Program
+            preprint narrative (Sections A, B, and C)
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            1915(b) Comprehensive (Capitated) Waiver Application Pre-print (Initial, Renewal,
+            Amendment)
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            State submission of the 1915(b) preprint narrative (Sections A, B, C and D) (non-FFS
+            Selective Contracting Waiver programs)
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            1915(b) Comprehensive (Capitated) Waiver Cost effectiveness spreadsheets (Initial,
+            Renewal, Amendment)
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            Appendix D Cost Effectiveness Demonstration for 1915(b) Waivers only (not applicable to
+            1915(b)(4) Fee-for-Service Selective Contracting programs)
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            1915(b)(4) FFS Selective Contracting (Streamlined) and 1915(b) Comprehensive (Capitated)
+            Waiver Independent Assessment (first two renewals only)
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            State submission of the findings from the Independent Assessment of their 1915(b) waiver
+            program
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">
+            Tribal Consultation (Initial, Renewal, Amendment)
+          </td>
+          <td className="border border-gray-300 px-4 py-2">
+            Document that outlines the changes the waiver action is making and the impact that
+            tribes can expect from the waiver action
+          </td>
+        </tr>
+        <tr>
+          <td className="border border-gray-300 px-4 py-2">Other</td>
+          <td className="border border-gray-300 px-4 py-2">
+            {`Any other documents or spreadsheets that are supplemental to
+                    the state's waiver application`}
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </>
+);
